Add copy to clipboard for hash in stake pool list
* reuse and rework PoolEntityContent
