+++ frontend/src/components/common/AddressesBreakdown.js
@@ -18,8 +18,7 @@ import WithModalState from '@/components/headless/modalState'
 import {useIsMobile} from '@/components/hooks/useBreakpoints'
 import {useI18n} from '@/i18n/helpers'
 import {routeTo} from '@/helpers/routes'
-import {AdaValue, Link} from '@/components/common'
-import CopyToClipboard from '@/components/common/CopyToClipboard'
+import {AdaValue, Link, CopyToClipboard} from '@/components/common'
 
 import type {Transaction} from '@/__generated__/schema.flow'
