+++ frontend/src/components/common/index.js
@@ -2,11 +2,14 @@
 
 export {default as AdaValue} from './AdaValue'
 export {default as ComparisonMatrix} from './ComparisonMatrix'
+export {default as CopyToClipboard} from './CopyToClipboard'
 export {default as DebugApolloError} from './DebugApolloError'
 export {default as EntityIdCard, EntityCardContent, EntityCardShell} from './EntityIdCard'
 export {default as ErrorOverlay} from './ErrorOverlay'
+export {default as FileInputHandler} from './FileInputHandler'
 export {default as Link} from './Link'
 export {default as LoadingError} from './LoadingError'
 export {default as LoadingOverlay} from './LoadingOverlay'
 export {default as MetricsCard} from './MetricsCard'
 export {default as Pagination} from './Pagination'
+export {default as PoolEntityContent} from './PoolEntityContent'
